module_4_deploy/pyproject.toml
[tool.poetry]
name = "04-deployment"
version = "0.1.0"
description = "This is for model inference"
authors = ["Manasa Thirumala <manasa_thirumala@epam.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
pandas = "^2.2.2"
flask = "^2.2.5"
mlflow = "^2.16.2"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
boto3 = "^1.35.21"
minio = "^7.2.8"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false